adding an example run

+++ tophat_populations/confusion_noise.py
@@ -30,7 +30,7 @@ def _newton_F(beta, gamma, alpha, x0=10.0, n_iter=12):
 def get_threshold_and_conf_noise_for_powerlaw(Tobs, delta_f_band, Sw, Ntot, alpha, rho_th, Amin,
                                               w_fdot, phase_variance=True, N_variance=False):
     Nbins = delta_f_band * Tobs # number of bins
-    sigma_0_sq = Sw / Tobs # time domain wn variance
+    sigma_0_sq = Sw / Tobs # time domain wn variance 
     mu = Ntot / Nbins # expected sources per bin
 
     beta = rho_th**2 * sigma_0_sq / Amin**2 / 4
@@ -51,7 +51,7 @@ def get_threshold_and_conf_noise_for_powerlaw(Tobs, delta_f_band, Sw, Ntot, alph
         # xth = bisec.run().params
     Ath = xth * Amin
  
-    Sconf = (4 * Tobs * Amin**2 / rho_th**2) * (xth**2 - beta)
+    Sconf = (4 * Tobs * Amin**2 / rho_th**2) * np.float64(xth**2 - beta)
     avgA2 = (alpha - 1) / (3 - alpha) * Amin**(alpha - 1) * (Ath**(3-alpha) - Amin**(3-alpha))
     avgA4 = (alpha - 1) / (5 - alpha) * Amin**(alpha - 1) * (Ath**(5-alpha) - Amin**(5-alpha))
     # amplitude variance
